fix: refactor helpers
