+++ src/utils/bonfida.ts
@@ -16,8 +16,8 @@ export const MINT_PREFIX = Buffer.from("tokenized_name");
 /**
  * This function can be used to retrieve the owner of a tokenized domain name
  *
- * @param connection The solana connection object to the RPC node
- * @param nameAccount The key of the domain name
+ * @param connection - The solana connection object to the RPC node
+ * @param nameAccount - The key of the domain name
  * @returns
  */
 export const retrieveNftOwner = async (connection: Connection, nameAccount: PublicKey) => {
@@ -50,7 +50,7 @@ export const retrieveNftOwner = async (connection: Connection, nameAccount: Publ
 /**
  * This function can be used to retrieve all the tokenized domains name
  *
- * @param connection The solana connection object to the RPC node
+ * @param connection - The solana connection object to the RPC node
  * @returns
  */
 // export const retrieveNfts = async (connection: Connection) => {
